+++ overrides-module.js
@@ -7,7 +7,7 @@ export default [
       'compat/compat': 'off',
       'no-console': 'off',
       'unicorn/no-anonymous-default-export': 'off',
-      'import/no-anonymous-default-export': 'off',
+      'import-x/no-anonymous-default-export': 'off',
       'n/no-process-env': 'off'
     }
   }