+++ standard.js
@@ -125,11 +125,11 @@ export default [{
     '@stylistic/wrap-iife': 'error',
     '@stylistic/yield-star-spacing': 'error',
     yoda: 'error',
-    'import/export': 'error',
-    'import/first': 'error',
-    'import/no-duplicates': 'error',
-    'import/no-named-default': 'error',
-    'import/no-webpack-loader-syntax': 'error'
+    'import-x/export': 'error',
+    'import-x/first': 'error',
+    'import-x/no-duplicates': 'error',
+    'import-x/no-named-default': 'error',
+    'import-x/no-webpack-loader-syntax': 'error'
   }
 }, {
   name: 'ash-nazg/standard/node',