+++ typings/eslint-plugin-import.d.ts
@@ -1,9 +1,0 @@
-declare module 'eslint-plugin-import' {
-  interface ConfigsStructure {
-    flatConfigs: {
-      recommended: import('eslint').Linter.Config
-    }
-  }
-  var importPlugin: ConfigsStructure
-  export = importPlugin;
-}