+++ CHANGES.md
@@ -1,5 +1,10 @@
 # eslint-config-ash-nazg CHANGES
 
+## 40.0.0
+
+- chore: update deps. and devDeps.
+- chore: switch to eslint-plugin-import-x
+
 ## 39.11.1
 
 - fix: ensure `SwitchCase: 0` is present on `@stylistic/indent` as seems no longer the default (at least when has object options)